requirements.txt
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
spectree==1.2.9
typing_extensions==4.12.0
urllib3==2.2.1
validate-docbr==1.10.0
Werkzeug==3.0.3
